Add detailed Extended Key Usage output
This adds a `go generate` script (also written in Go), to fetch the list
of Extended Key Usage OID's from the zmap/constants repo, and generate
the necessary EKU constants and structs. There is an additional script,
extended_key_usage_schema.sh, that generates the necessary SubRecord()
schema for EKU. Note that all of these need to be run manually when the
list of constants changes.