Remove NameConstraint checking from x509.Verify() (#21)
This updates x509.Verify() to no longer enforce name constraints when
passed an opts.DNSName. This was a haphazard at best implementation, and
did not fully enforce the connection anyway. NameConstraints are likely
best checked after chains have been built, and can be dependent on the
name being verified. They might also be enforced differently in
different stores. Therefore, it's probably best not to check them here
until we decide on how we want to enforce them.