Reland 1bafdfad8ce09a560d316b8b71335b5e1818f882 with fix
This relands the above commit, but fixes x509.ValidateWithStupidDetail
to not swap trusted/untrusted values due to a reversed error-checking if
statement.

----------

Return Current, Expired, and NeverValid chains (#20)

This updates x509.Verify() to return three different types of chains:
those that are currently valid, according to the time defined in the
VerifyOpts, chains that were valid at one time (e.g. are now expired),
and chains that were NeverValid at any time.