Reland 5a1d94c1219e5e7b12f6adecd4d47f102843a5a6 (#37)
This doesn't actually touch any code that runs during ZGrab.

------

Add initial zcrypto/verifier (#22)

Verifier is a higher-level module for doing in-depth certificate
verification, with the goal of being able to impersonate a browser or
specific root store's validation procedure. It uses the chaining code in
x509 to generate chains, and then apply's customizable verification on
top.

This includes a few "built-in" verifiers meant to represent the Firefox,
Microsoft, Apple, and Java trust stores, respectively.

The package is not yet complete, but should have roughly equivalent
functionality to x509.ValidateWithStupidDetail, which is now deprecated.