Add Graph model of PKI used by Verifier package (#85)
This introduces a Graph structure than can be used to model the PKI. It
implements a multigraph in which edges are certificates, and nodes are
(spki, subject) tuples. The head/source of the edge is the issuer, and the
tail/destination is the subject. Verifiers walk this graph to perform
certificate validation.