Clean up and remove old x509 code, condense files, investigate 1.10 migration (#155)
Version 1.10  of the standard crypto/x509 library introduces features which
zcrypto already supports independently, for example SANExtension parsing in
certificates.  Resolving these differences would risk breaking dependancies,
with little clear benefit, so it seems that diverging from the standard
crypto/x509 and pkix libaries is the most reasonable approach moving forward.

This PR condenses the zcrypto/x509 package to remove unused and old code. This
includes code for handling root stores, which should now be done with
zmap/rootfetch or with other means. It also removes the *_updated.go paradigm,
as it does not serve a clear purpose if we adopt a policy of longer strictly
follow updates to the standard libary.

Fixes #1