x509: support parsing CABF TorServiceDescriptors.
In CAB Forum Ballot 144[0] validation rules for `.onion` names were
added to the EV Guidelines. The rules were subsequently tweaked in
Ballot 198[1] (declared invalid) and Ballot 201[2]. The rules specify
that certificates with an `.onion` subject must include
a `TorServiceDescriptor` x509 extension with OID `2.23.140.1.31`. This
extension is an ASN.1 `SEQUENCE` of `TorServiceDescriptorHash` entries.

This branch adds support for recognizing and parsing the
`TorServiceDescriptor` extension in parsed `x509.Certificate` objects.
The individual `TorServiceDescriptorHash` objects are parsed and made
accessible on the parsed `x509.Certificate` as the
`TorServiceDescriptor` field, a slice of `x509.TorServiceDescriptorHash`
objects. The JSON marshaling of `x509.Certificate` objects is updated to
marshal this new field as `tor_service_descriptors`.

[0]: https://cabforum.org/2015/02/18/ballot-144-validation-rules-dot-onion-names/
[1]: https://cabforum.org/2017/05/08/ballot-198-onion-revisions/
[2]: https://cabforum.org/2017/06/08/2427/