x509: support parsing CABF TorServiceDescriptor extension. (#189)
* x509: support parsing CABF TorServiceDescriptors.

In CAB Forum Ballot 144[0] validation rules for `.onion` names were
added to the EV Guidelines. The rules were subsequently tweaked in
Ballot 198[1] (declared invalid) and Ballot 201[2]. The rules specify
that certificates with an `.onion` subject must include
a `TorServiceDescriptor` x509 extension with OID `2.23.140.1.31`. This
extension is an ASN.1 `SEQUENCE` of `TorServiceDescriptorHash` entries.

This branch adds support for recognizing and parsing the
`TorServiceDescriptor` extension in parsed `x509.Certificate` objects.
The individual `TorServiceDescriptorHash` objects are parsed and made
accessible on the parsed `x509.Certificate` as the
`TorServiceDescriptor` field, a slice of `x509.TorServiceDescriptorHash`
objects. The JSON marshaling of `x509.Certificate` objects is updated to
marshal this new field as `tor_service_descriptors`.

[0]: https://cabforum.org/2015/02/18/ballot-144-validation-rules-dot-onion-names/
[1]: https://cabforum.org/2017/05/08/ballot-198-onion-revisions/
[2]: https://cabforum.org/2017/06/08/2427/

* x509: be more lenient parsing TorServiceDescriptor ext.

* refactored errors to better match package convention. Using
  `asn1.SyntaxError` for structural parsing errors indicating
  misencoding.
* removed enforcement of known hash algorithm, stored friendly name in
  new `AlgorithmName` field.
* removed enforcement of bit length of hash byts vs hash algorithm.
* added field for raw `pkix.AlgorithmIdentifier`, `Algorithm`
* removed enforcement of valid UTF8 onion name field.

* x509: use snake_case for TorServideDescriptorHash fields

* nit: remove extra CR in TestTorServiceDescriptorHashJSON

* x509: rewrite `TestParseTorServiceDescriptorSyntax`.

The old `TestParseTorServiceDescriptorSyntax` implementation tested with
mock certificates. The mock certificates were opaque `.pem` files and
couldn't be easily reconstructed or understood.

The new `TestParseTorServiceDescriptorSyntax` implementation tests with
in-memory `pkix.Extension` instances specified per-testcase. This makes
the ASN.1 testdata for each extension available for inspection from
within the testcode.

* x509: split TorServiceDescHash code into separate files.

Also switch to `CertificateFingerprint` type vs `[]byte` for the
`TorServiceDescriptorHash.Hash` field so that it will be serialized as
hex and not base64.