deps: remove govalidator dep by adopting IsURL.
Since the MIT and Apache licenses are compatible[0] the `zcrypto`
project can avoid a larger dependency on the entire
`github.com/asaskevich/govalidator` library by simply adopting the
one function used: `IsURL`.

This commit places that function in `util/isURL.go` with the original
MIT license header, removes the dependency on `govalidator` and updates
the `zcrypto` LICENSE file to reference the MIT license for
`util/isURL.go` similar to the manner that the adopted Google code's
license is referenced.

IANAL but I believe this is all copacetic.

[0]: https://law.stackexchange.com/questions/6081/can-i-bundle-mit-licensed-components-in-a-apache-2-0-licensed-project