x509: add ExtensionsMap to Certificate/JSONCertificate.
A new `ExtensionsMap` field is added on the `x509.Certificate` object.
This field maps from the string form of a `pkix.Extension.Id` OID (e.g.
"1.3.6.1.5.5.7.1.1") to a raw `pkix.Extension` value. This lets
relying code check if a certificate has a given extension by OID in
O(1) time.

To accomplish this the `parseCertificate` function is updated to
populate this map at the same time as parsing the existing
`x509.Certificate.Extensions` field. The `x509.JSONCertificate` type and
`x509.Certificate.MarshalJSON` function are updated to persist the
`ExtensionsMap` field to JSON.