tls: Export SignatureAndHash
The SignatureAndHashes field in tls.Config is exported, but the type
signatureAndHashes is not, and neither are its fields.  Export this, and
name is SigAndHashes to avoid conflict with the exported type
SignatureAndHashes in tls_ka.go.

With this exported, callers can override the supported
SignatureAndHashes manually in their Config.