CI: automate PRs for weppos/publicsuffix-go updates. (#240)
The upstream `github.com/weppos/publicsuffix-go` data updates with
frequency. Since release tags are not published as often as the data
updates are merged it makes sense for ZCrypto to track upstream master.

Manually updating the dependency when `publicsuffix-go` is updated is
a drag. This commit adds a workflow that can open the update PRs
automatically similar to the way ZLint updates gTLD data.